+++ SalesWebMvc/SalesWebMvc/Models/SalesRecord.cs
@@ -8,7 +8,7 @@ namespace SalesWebMvc.Models
         // Informações do registro de venda
         public int Id { get; set; }            // Número de identificação
         public DateTime Date { get; set; }     // Data do registro da venda
-        public float SaleValue { get; set; }        // Valor da venda
+        public double SaleValue { get; set; }        // Valor da venda
         public SaleStatus Status { get; set; } // Estado atual da venda        
         public Seller Seller { get; set; } // Vendedor
 
@@ -19,7 +19,7 @@ namespace SalesWebMvc.Models
         }
 
         // Construtor #2
-        public SalesRecord(int id, DateTime date, float saleValue, SaleStatus status, Seller seller)
+        public SalesRecord(int id, DateTime date, double saleValue, SaleStatus status, Seller seller)
         {
             Id = id;
             Date = date;
